ExaBayes, version 1.3:
* threaded version now implemented with efficiency comparable to MPI
  version

* allreduce is implemented with many lockless queues and in
  asynchroneous communication trees

* chain swapping implemented with locked single producer, multiple
  consumer queue

+++ src/config/BlockRunParameters.hpp
@@ -53,7 +53,7 @@ private:
   double burninProportion; 
   int samplingFreq; 
   int numRunConv; 
-  int numGen; 
+  uint64_t numGen; 
   int numCoupledChains; 
   int printFreq; 
   double heatFactor ;
